Re-factored ExpireFIleTransfer to PurgeFileTransfer
