+++ src/Altinn.Broker.Application/MalwareScanResults/MalwareScanResultHandler.cs
@@ -2,7 +2,7 @@
 using System.Security.Claims;
 using System.Text.Json;
 
-using Altinn.Broker.Application.ExpireFileTransfer;
+using Altinn.Broker.Application.PurgeFileTransfer;
 using Altinn.Broker.Application.Middlewares;
 using Altinn.Broker.Core.Application;
 using Altinn.Broker.Core.Domain;
@@ -70,11 +70,10 @@ public class MalwareScanningResultHandler(
                 logger.LogWarning("Suspicious scan result for file transfer {fileTransferId} with body {body}", fileTransferId, JsonSerializer.Serialize(data));
                 await fileTransferStatusRepository.InsertFileTransferStatus(fileTransferId, Core.Domain.Enums.FileTransferStatus.Failed, $"Malware scan failed: {data.ScanResultType}. Extra details: " + JsonSerializer.Serialize(data.ScanResultDetails), cancellationToken);
                 backgroundJobClient.Enqueue(() => eventBus.Publish(AltinnEventType.UploadFailed, fileTransfer.ResourceId, fileTransferIdFromUri, fileTransfer.Sender.ActorExternalId, Guid.NewGuid()));
-                backgroundJobClient.Enqueue<ExpireFileTransferHandler>(handler => handler.Process(new ExpireFileTransferRequest
+                backgroundJobClient.Enqueue<PurgeFileTransferHandler>(handler => handler.Process(new PurgeFileTransferRequest
                 {
                     FileTransferId = fileTransfer.FileTransferId,
-                    Force = true,
-                    DoNotUpdateStatus = true
+                    PurgeTrigger = PurgeTrigger.MalwareScanFailed
                 }, null, CancellationToken.None));
             }
             return Task.CompletedTask;
